Backfill unit and integration tests
Taking advantage of a fake server to stub out the prismic api and
provide the most simple use case for it to do what it is intended to
do, download information from prismic and make it available to
middleman

For each test currently we do all of our work in an environment
variable. Since we need to inform at least one test what that
directory is, we pass that back via metadata. This isn't great, but
feels better than assigning an instance variable.

Configure circle to run with multiple versions of ruby, since this
should theoretically work with multiple versions. This only supports
the versions of ruby officially supported by ruby

* Add some basic unit tests
* Define integration happy path for prismic gem
* Use capybara_discoball and sinatra to stub out prismic
* Provide fake but realistic output from api
* Do everything in a temporary folder via an around hook
* Take advantage of <<~ to do some fancy indenting heredocs
* Use RUBY, makes it clearer what I'm writing and github will
  highlight it accordingling.
* Configure circleci