Add Middleman::Prismic::Preview middleware
Prismic supports a preview function by sending a browser to /preview
with a token and document_id on the query. This middleware is intended
to hook into Middleman to add a /preview endpoint that will fetch the
preview content by token and display the content in the Middleman
server.

In your Middleman app, add the following to `config.rb` to activate
this middleware:

```ruby
require "middleman-prismic/preview"

use Middleman::Prismic::Preview
```

Now when you run `middleman`, any requests to `/preview` will be
intercepted. The `token` param will be read and used to execute
`middleman prismic --ref ${token}`. Because Middleman is running as a
server, when the content from Prismic is downloaded, it will show the
live content when browsing the Middleman site.

Next piece of work will be to support `document_id` by allowing a
Prismic Link Resolver into the middleware when called `use`.